Alter 'payment' table
Now understand that we can ignore 'BILL' column. Remove this condition for filtering so that we only retain non-zero payments. This will be essential for changes to `class_student` table and `ClassInfoFrame` class roll logic, as students will be listed as members of the class regardless of whether they have paid. Their names will only be displayed within the program upon payment