Alter `class_student` table
Before, `class_student` would only contain students that were present in `clsbymon.dbf`. The old program works so that students in `clsbymon.dbf` are those who are paid for the current month.

Now, change table to include all students who are linked to a class, regardless of payment. Then `ClassInfoFrame` will not show all students for the relevant class, but rather display only the students in that class which are paid for current month