Create logic for billing
Users toggle 'bill' status for a given student/month/year by simply clicking the relevant row in the payment_frame. When toggled ON, an asterisk (*) appears as in the old program (`toggle_bill()`), and then the database is updated (`bill_student()`) so that 1) a new bill record is created in the `bill` table and 2) an asterisk (*) is added to the relevant column in `STUD00.dbf` so that the old program remains consistent.

Vice versa also applies (toggling bill OFF will remove asterisk, delete relevant record from `bill` table, and remove asterisk (*) from relevant column in `STUD00.dbf`)