Add `do_not_load` option, modify `edit_info()`, create `validate_entryboxes()`
Take data validation logic from `edit_info()` and move into a separate function `validate_entryboxes()` for better organization and for easier use in the NewStudentDialog class.

`do_not_load` option given to `transform_to_rdb()` as a list of table names which we do not want to load (transform) from the original DBF files. For now, just `note` is in this list, as the new database represents class/student notes as unlimited-length strings, and therefore these would be overwritten if we load from DBF every time the program starts.