Error handling / messaging in `NewStudentDialog`
Implement extra handling to prevent errors. Data validation was already occurring, but now program will pause and warn user when 1) first or last name is missing or 2) the user tries to exit the window. In both scenarios, the window will prompt the user to take some kind of action (either fixing the names or confirming that they indeed wish to close the window and cancel the 'new student' operation)