Begin implementing SQLite database
Before, we transformed the DBF files into a relational database structure but represented those files as Pandas DataFrames and CSV files. On this branch, we will take this same representation but use an actual database connection via SQLite.

Keep `StudentDatabase` class for now as a way to handle modifying DBF and SQLite databases in parallel; eventually we may want to restructure this class

Modify `functions.transform_to_rdb` to have an option to save out to `.db` after transformation rather than `csv`