Replace `StudentDatabase` table attributes with SQLite
Now rather than storing tables as attributes of `StudentDatabase` object, we have a database connection which we can query.

There will be no data stored in DataFrames at runtime, rather all data displayed in the program will be queried and modified in real-time using the SQLite database.

This commit represents a large shift of practically every bit of code in the program from DataFrames to a database connection. There may still be instances that need to be cleaned up in future commits, but at least the majority of the logic has been re-written here.